requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
Pillow==10.0.0
pydantic==2.5.0
